api_requirements.txt
flask==3.0.3
ultralytics==8.3.107
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
timm==0.9.12
gunicorn==21.2.0
Pillow==10.4.0 